Restructure Validation files
